+++ dukeLogger.log
@@ -1,8 +1,6 @@
-Mar 21, 2020 9:15:50 AM seedu.duke.Duke run
+Mar 21, 2020 2:38:11 PM seedu.duke.Duke run
 INFO: Application started Execution
-Mar 21, 2020 9:15:52 AM seedu.duke.Duke run
-INFO: Application transitioning to task mode
-Mar 21, 2020 9:16:30 AM seedu.duke.Duke run
+Mar 21, 2020 2:38:14 PM seedu.duke.Duke run
 INFO: Application is going to exit
-Mar 21, 2020 9:16:30 AM seedu.duke.Duke run
+Mar 21, 2020 2:38:14 PM seedu.duke.Duke run
 INFO: Application closed successfully
