+++ src/test/java/task/TaskListTest.java
@@ -1,8 +1,8 @@
-package task.event;
+package task;
 
 import org.junit.jupiter.api.Test;
-import task.Event;
 import ui.Ui;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
