+++ text-ui-test/EXPECTED.TXT
@@ -254,7 +254,7 @@ ________________________________________________________________________________
 	 Ground Level
 	 Faculty: Utown
 	 Port: true
-	 Indoor: false
+	 Indoor: true
 	 Maximum number of Pax: 5
 	 __________________________________________________________
 	 __________________________________________________________
@@ -334,7 +334,7 @@ ________________________________________________________________________________
 	 Ground Level
 	 Faculty: Utown
 	 Port: true
-	 Indoor: false
+	 Indoor: true
 	 Maximum number of Pax: 5
 	 __________________________________________________________
 	 Please enter the location for your desired study area.
@@ -347,6 +347,15 @@ ________________________________________________________________________________
 ________________________________________________________________________________________________
 ________________________________________________________________________________________________
 	 Here are the available study areas!
+	 __________________________________________________________
+	 Name: Starbucks
+	 Address: 2 College Avenue West Education Resource Centre,
+	 Ground Level
+	 Faculty: Utown
+	 Port: true
+	 Indoor: true
+	 Maximum number of Pax: 5
+	 __________________________________________________________
 	 __________________________________________________________
 	 Name: MacCommons
 	 Address: 2 College Avenue West Education Resource Centre,
@@ -467,7 +476,7 @@ ________________________________________________________________________________
 	 Ground Level
 	 Faculty: Utown
 	 Port: true
-	 Indoor: false
+	 Indoor: true
 	 Maximum number of Pax: 5
 	 __________________________________________________________
 	 __________________________________________________________
@@ -628,7 +637,7 @@ ________________________________________________________________________________
 	 Ground Level
 	 Faculty: Utown
 	 Port: true
-	 Indoor: false
+	 Indoor: true
 	 Maximum number of Pax: 5
 	 __________________________________________________________
 	 __________________________________________________________
