+++ text-ui-test/dukeLogger.log
@@ -1,36 +1,36 @@
-Mar 21, 2020 9:24:06 AM seedu.duke.Duke run
+Mar 21, 2020 9:43:37 AM seedu.duke.Duke run
 INFO: Application started Execution
-Mar 21, 2020 9:24:06 AM seedu.duke.Duke run
+Mar 21, 2020 9:43:37 AM seedu.duke.Duke run
 INFO: Application transitioning to task mode
-Mar 21, 2020 9:24:06 AM task.TaskList deleteTask
+Mar 21, 2020 9:43:37 AM task.TaskList deleteTask
 INFO: Index specified to be deleted is out of bounds
-Mar 21, 2020 9:24:06 AM task.Event getFieldToBeEdited
+Mar 21, 2020 9:43:37 AM task.Event getFieldToBeEdited
 INFO: Invalid option entered by user
-Mar 21, 2020 9:24:06 AM task.Event getFieldToBeEdited
+Mar 21, 2020 9:43:37 AM task.Event getFieldToBeEdited
 INFO: Invalid option entered by user
-Mar 21, 2020 9:24:06 AM task.Event parseDate
+Mar 21, 2020 9:43:37 AM task.Event parseDate
 INFO: Date after current date exception thrown
-Mar 21, 2020 9:24:06 AM task.Event editDate
+Mar 21, 2020 9:43:37 AM task.Event editDate
 INFO: Invalid date entered by user
-Mar 21, 2020 9:24:06 AM task.Deadline getFieldToBeEdited
+Mar 21, 2020 9:43:37 AM task.Deadline getFieldToBeEdited
 INFO: Invalid option entered by user
-Mar 21, 2020 9:24:06 AM seedu.duke.Duke run
+Mar 21, 2020 9:43:37 AM seedu.duke.Duke run
 INFO: Application transitioning to task mode
-Mar 21, 2020 9:24:06 AM seedu.duke.Duke run
+Mar 21, 2020 9:43:37 AM seedu.duke.Duke run
 INFO: Application transitioning to study area mode
-Mar 21, 2020 9:24:06 AM studyarea.StudyAreaList checkDuplicate
+Mar 21, 2020 9:43:37 AM studyarea.StudyAreaList checkDuplicate
 INFO: Duplicate flags entered by user
-Mar 21, 2020 9:24:06 AM studyarea.StudyAreaList checkDuplicate
+Mar 21, 2020 9:43:37 AM studyarea.StudyAreaList checkDuplicate
 INFO: Duplicate flags entered by user
-Mar 21, 2020 9:24:06 AM studyarea.StudyAreaList checkInteger
+Mar 21, 2020 9:43:37 AM studyarea.StudyAreaList checkInteger
 INFO: no string exist after "-s"
-Mar 21, 2020 9:24:06 AM studyarea.StudyAreaList checkInteger
+Mar 21, 2020 9:43:37 AM studyarea.StudyAreaList checkInteger
 INFO: User entered an invalid number for integer input
-Mar 21, 2020 9:24:06 AM studyarea.StudyAreaList checkOnlyFlag
+Mar 21, 2020 9:43:37 AM studyarea.StudyAreaList checkOnlyFlag
 INFO: Identifier not entered with flag by user
-Mar 21, 2020 9:24:06 AM studyarea.StudyAreaList checkFlag
+Mar 21, 2020 9:43:37 AM studyarea.StudyAreaList checkFlag
 INFO: Wrong flag used by user while searching for study areas
-Mar 21, 2020 9:24:06 AM seedu.duke.Duke run
+Mar 21, 2020 9:43:37 AM seedu.duke.Duke run
 INFO: Application is going to exit
-Mar 21, 2020 9:24:06 AM seedu.duke.Duke run
+Mar 21, 2020 9:43:37 AM seedu.duke.Duke run
 INFO: Application closed successfully
