+++ text-ui-test/library/dictionary.txt
@@ -1,0 +1,10 @@
+engineering~eng~engin~engine~foe
+utown~university town~u-town~u town
+maccommons~mccommons~mac commons
+education resource centre~erc
+stephen riady centre~stephen~riady~src
+ocbc~frank~frank by ocbc
+starbucks~bucks~bux~sbux
+pc commons~pccommons~pccommon
+yih~yusof ishak~yusof~ishak
+supersnacks~super snacks~super snack
