+++ text-ui-test/library/taskList.txt
@@ -1,0 +1,2 @@
+D#assignment#2020-05-02#14:00#2
+E#team meeting#2020-06-06#12:00#13:00#1
